GUI support
A GUI app in Docker is less obvious.

Document how to run the container, by sharing the X11 socket:
http://fabiorehm.com/blog/2014/09/11/running-gui-apps-with-docker/

And install the Basic Desktop group, to have all the required libraries
inside the container. Normally only the bare minimum should be
installed, but I'll see later what is really necessary. I prefer to have
something working first (and in fact, the GUI can now be launched).