Reorganize dnf commands
First, install the basic desktop group (it's quite huge to install).
Then, install the dependencies of Pupil. That latter list of packages is
more subject to changes.

And merge two dnf commands. The second dnf command needed to re-download
all the metadata, since there was the 'dnf clean all' on the previous
RUN command (which is a best-practice to reduce image sizes).

And s/update/upgrade/. "update" was AFAIK for backward compat with yum.
The manpage of dnf refers to "upgrade".